creatTournee ok, il reste creatLivraison dans TourneeComponent

+++ rest-api/src/main/java/fr/uga/l3miage/integrator/requests/LivraisonCreationRequest.java
@@ -11,7 +11,7 @@ import lombok.Data;
 public class LivraisonCreationRequest {
     private String reference;
     private String etat;
-    private Float montant;
+    private double montant;
 
 
 }
